Add new classes v1
